Restructured for future reuse
