+++ src/BotPlutusInterface/Config/Cardano/Api/Shelley.hs
@@ -1,0 +1,260 @@
+{-# LANGUAGE ApplicativeDo #-}
+{-# LANGUAGE RecordWildCards #-}
+
+{-# OPTIONS -fno-warn-orphans  #-}
+
+module BotPlutusInterface.Config.Cardano.Api.Shelley () where
+
+import BotPlutusInterface.Config.Base
+import BotPlutusInterface.Config.Cardano.Api ()
+import BotPlutusInterface.Config.Types
+import Cardano.Api (EpochNo (EpochNo))
+import Cardano.Api.Shelley (ProtocolParameters (..))
+import Config
+import Config.Schema
+import Data.Text qualified as Text
+import Numeric.Natural (Natural)
+import Prelude
+
+instance ToValue (Natural, Natural) where
+  toValue (a, b) = List () [toValue a, toValue b]
+
+protocolVersionSpec :: ValueSpec (Natural, Natural)
+protocolVersionSpec =
+  customSpec
+    "protocol version"
+    (listSpec naturalSpec)
+    ( \case
+        [major, minor] -> Right (major, minor)
+        _ -> Left $ Text.pack "[MAJOR, MINOR]"
+    )
+
+{- ORMOLU_DISABLE -}
+instance ToValue ProtocolParameters where
+  toValue
+    ( ProtocolParameters
+        protocolParamProtocolVersion
+        protocolParamDecentralization
+        protocolParamExtraPraosEntropy
+        protocolParamMaxBlockHeaderSize
+        protocolParamMaxBlockBodySize
+        protocolParamMaxTxSize
+        protocolParamTxFeeFixed
+        protocolParamTxFeePerByte
+        protocolParamMinUTxOValue
+        protocolParamStakeAddressDeposit
+        protocolParamStakePoolDeposit
+        protocolParamMinPoolCost
+        protocolParamPoolRetireMaxEpoch
+        protocolParamStakePoolTargetNum
+        protocolParamPoolPledgeInfluence
+        protocolParamMonetaryExpansion
+        protocolParamTreasuryCut
+        protocolParamUTxOCostPerWord
+        protocolParamCostModels
+        protocolParamPrices
+        protocolParamMaxTxExUnits
+        protocolParamMaxBlockExUnits
+        protocolParamMaxValueSize
+        protocolParamCollateralPercent
+        protocolParamMaxCollateralInputs
+      ) =
+      Sections
+        ()
+        [ Section () "protocolVersion"     $ toValue protocolParamProtocolVersion
+        , Section () "decentralization"    $ toValue protocolParamDecentralization
+        , Section () "extraPraosEntropy"   $ toValue protocolParamExtraPraosEntropy
+        , Section () "maxBlockHeaderSize"  $ toValue protocolParamMaxBlockHeaderSize
+        , Section () "maxBlockBodySize"    $ toValue protocolParamMaxBlockBodySize
+        , Section () "maxTxSize"           $ toValue protocolParamMaxTxSize
+        , Section () "txFeeFixed"          $ toValue protocolParamTxFeeFixed
+        , Section () "txFeePerByte"        $ toValue protocolParamTxFeePerByte
+        , Section () "minUTxOValue"        $ toValue protocolParamMinUTxOValue
+        , Section () "stakeAddressDeposit" $ toValue protocolParamStakeAddressDeposit
+        , Section () "stakePoolDeposit"    $ toValue protocolParamStakePoolDeposit
+        , Section () "minPoolCost"         $ toValue protocolParamMinPoolCost
+        , Section () "poolRetireMaxEpoch"  $ toValue protocolParamPoolRetireMaxEpoch
+        , Section () "stakePoolTargetNum"  $ toValue protocolParamStakePoolTargetNum
+        , Section () "poolPledgeInfluence" $ toValue protocolParamPoolPledgeInfluence
+        , Section () "monetaryExpansion"   $ toValue protocolParamMonetaryExpansion
+        , Section () "treasuryCut"         $ toValue protocolParamTreasuryCut
+        , Section () "UTxOCostPerWord"     $ toValue protocolParamUTxOCostPerWord
+        , Section () "costModels"          $ toValue protocolParamCostModels
+        , Section () "prices"              $ toValue protocolParamPrices
+        , Section () "maxTxExUnits"        $ toValue protocolParamMaxTxExUnits
+        , Section () "maxBlockExUnits"     $ toValue protocolParamMaxBlockExUnits
+        , Section () "maxValueSize"        $ toValue protocolParamMaxValueSize
+        , Section () "collateralPercent"   $ toValue protocolParamCollateralPercent
+        , Section () "maxCollateralInputs" $ toValue protocolParamMaxCollateralInputs
+        ]
+{- ORMOLU_ENABLE -}
+
+instance HasSpec ProtocolParameters where
+  anySpec = sectionsSpec "ProtocolParameters" $ do
+    protocolParamProtocolVersion <-
+      optSectionFromDef'
+        protocolParamProtocolVersion
+        "protocolVersion"
+        protocolVersionSpec
+        "Protocol version, major and minor. Updating the major version is used to trigger hard forks."
+
+    protocolParamDecentralization <-
+      optSectionFromDef'
+        protocolParamDecentralization
+        "decentralization"
+        customRationalSpec
+        "The decentralization parameter. This is fraction of slots that belong to the BFT overlay schedule, rather than the Praos schedule. So 1 means fully centralised, while 0 means fully decentralised."
+
+    protocolParamExtraPraosEntropy <-
+      optSectionFromDef'
+        protocolParamExtraPraosEntropy
+        "extraPraosEntropy"
+        (maybeSpec anySpec)
+        "Extra entropy for the Praos per-epoch nonce."
+
+    protocolParamMaxBlockHeaderSize <-
+      optSectionFromDef
+        protocolParamMaxBlockHeaderSize
+        "maxBlockHeaderSize"
+        "The maximum permitted size of a block header."
+
+    protocolParamMaxBlockBodySize <-
+      optSectionFromDef
+        protocolParamMaxBlockBodySize
+        "maxBlockBodySize"
+        "The maximum permitted size of the block body (that is, the block payload, without the block header)."
+
+    protocolParamMaxTxSize <-
+      optSectionFromDef
+        protocolParamMaxTxSize
+        "maxTxSize"
+        "The maximum permitted size of the block body (that is, the block payload, without the block header)."
+
+    protocolParamTxFeeFixed <-
+      optSectionFromDef
+        protocolParamTxFeeFixed
+        "txFeeFixed"
+        "The constant factor for the minimum fee calculation."
+
+    protocolParamTxFeePerByte <-
+      optSectionFromDef
+        protocolParamTxFeePerByte
+        "txFeePerByte"
+        "The linear factor for the minimum fee calculation."
+
+    protocolParamMinUTxOValue <-
+      optSectionFromDef'
+        protocolParamMinUTxOValue
+        "minUTxOValue"
+        (maybeSpec anySpec)
+        "The minimum permitted value for new UTxO entries, ie for transaction outputs."
+
+    protocolParamStakePoolDeposit <-
+      optSectionFromDef
+        protocolParamStakePoolDeposit
+        "stakePoolDeposit"
+        "The deposit required to register a stake pool."
+
+    protocolParamStakeAddressDeposit <-
+      optSectionFromDef
+        protocolParamStakeAddressDeposit
+        "stakeAddressDeposit"
+        "The deposit required to register a stake address."
+
+    protocolParamMinPoolCost <-
+      optSectionFromDef
+        protocolParamMinPoolCost
+        "minPoolCost"
+        "The minimum value that stake pools are permitted to declare for their cost parameter."
+
+    protocolParamPoolRetireMaxEpoch <-
+      optSectionFromDef'
+        protocolParamPoolRetireMaxEpoch
+        "poolRetireMaxEpoch"
+        (EpochNo <$> anySpec)
+        "The maximum number of epochs into the future that stake pools are permitted to schedule a retirement."
+
+    protocolParamStakePoolTargetNum <-
+      optSectionFromDef
+        protocolParamStakePoolTargetNum
+        "stakePoolTargetNum"
+        "The equilibrium target number of stake pools."
+
+    protocolParamPoolPledgeInfluence <-
+      optSectionFromDef'
+        protocolParamPoolPledgeInfluence
+        "poolPledgeInfluence"
+        customRationalSpec
+        "The influence of the pledge in stake pool rewards."
+
+    protocolParamMonetaryExpansion <-
+      optSectionFromDef'
+        protocolParamMonetaryExpansion
+        "monetaryExpansion"
+        customRationalSpec
+        "The monetary expansion rate. This determines the fraction of the reserves that are added to the fee pot each epoch."
+
+    protocolParamTreasuryCut <-
+      optSectionFromDef'
+        protocolParamTreasuryCut
+        "treasuryCut"
+        customRationalSpec
+        "The fraction of the fee pot each epoch that goes to the treasury."
+
+    protocolParamUTxOCostPerWord <-
+      optSectionFromDef'
+        protocolParamUTxOCostPerWord
+        "UTxOCostPerWord"
+        (maybeSpec anySpec)
+        "Cost in ada per word of UTxO storage."
+
+    protocolParamCostModels <-
+      optSectionFromDef'
+        protocolParamCostModels
+        "costModels"
+        anySpec
+        "Cost models for script languages that use them."
+
+    protocolParamPrices <-
+      optSectionFromDef'
+        protocolParamPrices
+        "prices"
+        (maybeSpec anySpec)
+        "Price of execution units for script languages that use them."
+
+    protocolParamMaxTxExUnits <-
+      optSectionFromDef'
+        protocolParamMaxTxExUnits
+        "maxTxExUnits"
+        (maybeSpec anySpec)
+        "Max total script execution resources units allowed per tx."
+
+    protocolParamMaxBlockExUnits <-
+      optSectionFromDef'
+        protocolParamMaxBlockExUnits
+        "maxBlockExUnits"
+        (maybeSpec anySpec)
+        "Max size of a Value in a tx output."
+
+    protocolParamMaxValueSize <-
+      optSectionFromDef'
+        protocolParamMaxValueSize
+        "maxValueSize"
+        (maybeSpec naturalSpec)
+        "Max size of a Value in a tx output."
+
+    protocolParamCollateralPercent <-
+      optSectionFromDef'
+        protocolParamCollateralPercent
+        "collateralPercent"
+        (maybeSpec naturalSpec)
+        "The percentage of the script contribution to the txfee that must be provided as collateral inputs when including Plutus scripts."
+
+    protocolParamMaxCollateralInputs <-
+      optSectionFromDef'
+        protocolParamMaxCollateralInputs
+        "maxCollateralInputs"
+        (maybeSpec naturalSpec)
+        "The maximum number of collateral inputs allowed in a transaction."
+
+    pure ProtocolParameters {..}
